Reduce code duplication in R benefits calculator
Now it uses an array of values (with an explaining comment) which should
reduce code duplication by iterating through it. The only thing that the
new R powers need to add is a value in this array (which contains two
functions, one for bonus calc and one for text information).

Unlocks are still in the same old switch-case block.

Fix some things I broke :P